refactor(date): move from experimental into main components directory
Moves component into the main `src/components` from `__experimental__`

BREAKING CHANGE: Date has been moved to src/components/date. There is a codemod available to assist
with this upgrade
(https://github.com/Sage/carbon-codemod/tree/master/transforms/move-experimental-components)
