+++ cypress/features/regression/accessibility/accessibilityExperimental.feature
@@ -8,7 +8,6 @@ Feature: Accessibility tests - Experimental folder
     Examples:
       | component           |
       | date-range          |
-      | date-input          |
       | decimal-input       |
       | groupedcharacter    |
       | number-input        |
@@ -22,7 +21,6 @@ Feature: Accessibility tests - Experimental folder
     Examples:
       | component           |
       | date-range          |
-      | date-input          |
       | decimal-input       |
       | groupedcharacter    |
       | simple-color-picker |
@@ -35,7 +33,6 @@ Feature: Accessibility tests - Experimental folder
     Examples:
       | component           |
       | date-range          |
-      | date-input          |
       | decimal-input       |
       | groupedcharacter    |
       | simple-color-picker |
@@ -48,7 +45,6 @@ Feature: Accessibility tests - Experimental folder
     Examples:
       | component           |
       | date-range          |
-      | date-input          |
       | decimal-input       |
       | groupedcharacter    |
       | simple-color-picker |
