+++ src/__experimental__/components/date-range/date-range.component.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import styledSystemPropTypes from "@styled-system/prop-types";
 
-import DateInput from "../date";
+import DateInput from "../../../components/date";
 import { filterStyledSystemMarginProps } from "../../../style/utils";
 import tagComponent from "../../../utils/helpers/tags";
 import StyledDateRange from "./date-range.style";
