+++ src/components/date/date.stories.mdx
@@ -1,14 +1,14 @@
 import { useState } from "react";
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 
-import StyledSystemProps from "../../../../.storybook/utils/styled-system-props";
+import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 import DateInput, { BaseDateInput } from ".";
-import { OriginalTextbox } from "../../../components/textbox";
-import Button from "../../../components/button";
-import Box from "../../../components/box";
+import { OriginalTextbox } from "../textbox";
+import Button from "../button";
+import Box from "../box";
 
 <Meta
-  title="Experimental/Date Input"
+  title="Design System/Date Input"
   parameters={{ info: { disable: true } }}
 />
 
@@ -27,7 +27,7 @@ import Box from "../../../components/box";
 ## Quick Start
 
 ```javascript
-import DateInput from "carbon-react/lib/__experimental__/components/date";
+import DateInput from "carbon-react/lib/components/date";
 ```
 
 ## Examples
