+++ src/components/date/day-picker.style.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import baseTheme from "../../../style/themes/base";
+import baseTheme from "../../style/themes/base";
 
 // Styles copied from https://github.com/gpbl/react-day-picker/blob/v6.1.1/src/style.css
 const addReactDayPickerStyles = () => `
