+++ src/components/date/navbar/navbar.component.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import StyledButton from "./button.style";
 import StyledNavbar from "./navbar.style";
-import Icon from "../../../../components/icon";
+import Icon from "../../icon";
 
 const Navbar = ({ onPreviousClick, onNextClick, ...props }) => {
   return (
