+++ src/components/date/navbar/navbar.style.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import baseTheme from "../../../../style/themes/base";
+import baseTheme from "../../../style/themes/base";
 
 const StyledNavbar = styled.div`
   &.DayPicker-NavBar {
