+++ src/components/date/weekday/weekday.component.js
