+++ src/components/date/weekday/weekday.spec.js
