+++ src/components/date/weekday/weekday.style.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import baseTheme from "../../../../style/themes/base";
+import baseTheme from "../../../style/themes/base";
 
 const StyledWeekday = styled.div`
   &,
