+++ src/components/dialog/dialog.stories.js
@@ -6,7 +6,7 @@ import Dialog from "./dialog.component";
 import Form from "../form";
 import Textbox from "../textbox";
 import Button from "../button";
-import DateInput from "../../__experimental__/components/date";
+import DateInput from "../date";
 import { Checkbox } from "../checkbox";
 import { Select, Option } from "../select";
