+++ src/components/form/form.stories.mdx
@@ -8,7 +8,7 @@ import {
   RadioButton,
   RadioButtonGroup,
 } from "../../__experimental__/components/radio-button";
-import DateInput from "../../__experimental__/components/date";
+import DateInput from "../date";
 import { Select, Option } from "../select";
 import { Checkbox } from "../checkbox";
 import Hr from "../../components/hr";
